+++ MatchMakingLobby/MatchSettings/Joust.php
@@ -13,7 +13,7 @@ class Joust implements MatchSettings
 	public function getLobbyScriptSettings()
 	{
 		$rules = array(
-			'S_UseLobby' => 1,
+			'S_UseLobby' => true,
 			'S_LobbyTimePerMap' => 1800
 		);
 		return $rules;
@@ -22,7 +22,7 @@ class Joust implements MatchSettings
 	public function getMatchScriptSettings()
 	{
 		$rules = array(
-			'S_UseLobby' => 0,
+			'S_UseLobby' => false,
 			'S_UseWarmup' => true
 		);
 		return $rules;